test(infra): share social, chat, and channel fixtures
Reduce repeated test scaffolding across social handlers, chat connectors, and dock coverage so follow-on fixes can reuse stable shared fixtures instead of re-implementing mocks.
