+++ tests/connectors/chat/test_base.py
@@ -14,7 +14,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from aragora.connectors.chat.base import ChatPlatformConnector
+from .conftest import StubConnector
 from aragora.connectors.chat.models import (
     BotCommand,
     FileAttachment,
@@ -22,98 +22,6 @@ from aragora.connectors.chat.models import (
     UserInteraction,
     WebhookEvent,
 )
-
-
-# ============================================================================
-# Concrete Test Subclass
-# ============================================================================
-
-
-class StubConnector(ChatPlatformConnector):
-    """Minimal concrete implementation for testing base class behaviour."""
-
-    @property
-    def platform_name(self) -> str:
-        return "stub"
-
-    @property
-    def platform_display_name(self) -> str:
-        return "Stub Platform"
-
-    async def send_message(self, channel_id, text, blocks=None, thread_id=None, **kw):
-        return SendMessageResponse(success=True, message_id="msg-1")
-
-    async def update_message(self, channel_id, message_id, text, blocks=None, **kw):
-        return SendMessageResponse(success=True, message_id=message_id)
-
-    async def delete_message(self, channel_id, message_id, **kw):
-        return True
-
-    async def respond_to_command(self, command, text, blocks=None, ephemeral=True, **kw):
-        return SendMessageResponse(success=True)
-
-    async def respond_to_interaction(
-        self, interaction, text, blocks=None, replace_original=False, **kw
-    ):
-        return SendMessageResponse(success=True)
-
-    async def upload_file(
-        self,
-        channel_id,
-        content,
-        filename,
-        content_type="application/octet-stream",
-        title=None,
-        thread_id=None,
-        **kw,
-    ):
-        return FileAttachment(
-            id="file-1", filename=filename, content_type=content_type, size=len(content)
-        )
-
-    async def download_file(self, file_id, **kw):
-        return FileAttachment(id=file_id, filename="test.txt", content_type="text/plain", size=0)
-
-    def format_blocks(self, title=None, body=None, fields=None, actions=None, **kw):
-        return [{"type": "section", "text": body}]
-
-    def format_button(self, text, action_id, value=None, style="default", url=None):
-        return {"type": "button", "text": text, "action_id": action_id}
-
-    def verify_webhook(self, headers, body):
-        return True
-
-    def parse_webhook_event(self, headers, body):
-        return WebhookEvent(platform="stub", event_type="message")
-
-
-# ============================================================================
-# Fixtures
-# ============================================================================
-
-
-@pytest.fixture(autouse=True)
-def reset_circuit_breakers():
-    """Clear all circuit breakers before each test to ensure isolation."""
-    from aragora.resilience import _circuit_breakers, _circuit_breakers_lock
-
-    with _circuit_breakers_lock:
-        _circuit_breakers.clear()
-    yield
-    with _circuit_breakers_lock:
-        _circuit_breakers.clear()
-
-
-@pytest.fixture
-def connector():
-    """Create a default StubConnector with circuit breaker enabled."""
-    return StubConnector(bot_token="tok-123", signing_secret="sec-456")
-
-
-@pytest.fixture
-def connector_no_cb():
-    """Create a StubConnector with circuit breaker disabled."""
-    return StubConnector(bot_token="tok-123", enable_circuit_breaker=False)
 
 
 # ============================================================================
