+++ tests/server/handlers/social/test_social_media.py
@@ -1,93 +1,16 @@
 """Tests for aragora.server.handlers.social.social_media - Social Media Handler."""
 
-import sys
-import types as _types_mod
-
-# Pre-stub Slack modules to prevent import chain failures
-_SLACK_ATTRS = [
-    "SlackHandler",
-    "get_slack_handler",
-    "get_slack_integration",
-    "get_workspace_store",
-    "resolve_workspace",
-    "create_tracked_task",
-    "_validate_slack_url",
-    "SLACK_SIGNING_SECRET",
-    "SLACK_BOT_TOKEN",
-    "SLACK_WEBHOOK_URL",
-    "SLACK_ALLOWED_DOMAINS",
-    "SignatureVerifierMixin",
-    "CommandsMixin",
-    "EventsMixin",
-    "init_slack_handler",
-]
-for _mod_name in (
-    "aragora.server.handlers.social.slack.handler",
-    "aragora.server.handlers.social.slack",
-    "aragora.server.handlers.social._slack_impl",
-):
-    if _mod_name not in sys.modules:
-        _m = _types_mod.ModuleType(_mod_name)
-        for _a in _SLACK_ATTRS:
-            setattr(_m, _a, None)
-        sys.modules[_mod_name] = _m
-
-import json
-from dataclasses import dataclass
-from io import BytesIO
 from pathlib import Path
-from typing import Any, Optional
+from typing import Any
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
+from .conftest import MockHandler, install_social_slack_stubs
+
+install_social_slack_stubs()
+
 from aragora.server.handlers.social.social_media import SocialMediaHandler
-
-
-# ===========================================================================
-# Mock Classes
-# ===========================================================================
-
-
-@dataclass
-class MockUser:
-    """Mock authenticated user."""
-
-    user_id: str = "user-123"
-    id: str = "user-123"
-    org_id: str = "org-123"
-    email: str = "test@example.com"
-    name: str = "Test User"
-
-
-class MockHandler:
-    """Mock HTTP request handler."""
-
-    def __init__(
-        self,
-        body: bytes = b"",
-        headers: dict[str, str] | None = None,
-        path: str = "/",
-        method: str = "GET",
-    ):
-        self._body = body
-        self.headers = headers or {"Content-Length": str(len(body))}
-        self.path = path
-        self.command = method
-        self.rfile = BytesIO(body)
-        self.client_address = ("127.0.0.1", 12345)
-
-    @classmethod
-    def with_json_body(cls, data: dict[str, Any], **kwargs) -> "MockHandler":
-        body = json.dumps(data).encode("utf-8")
-        headers = {
-            "Content-Type": "application/json",
-            "Content-Length": str(len(body)),
-        }
-        return cls(body=body, headers=headers, **kwargs)
-
-    def get_argument(self, name: str, default: str = None) -> str | None:
-        return default
 
 
 # ===========================================================================
@@ -105,18 +28,6 @@ def reset_oauth_state():
     except Exception:
         pass
     yield
-
-
-@pytest.fixture
-def mock_user():
-    return MockUser()
-
-
-@pytest.fixture
-def mock_user_store(mock_user):
-    store = MagicMock()
-    store.get_user_by_id.return_value = mock_user
-    return store
 
 
 @pytest.fixture
@@ -181,15 +92,20 @@ def mock_audio_store():
 
 @pytest.fixture
 def handler_context(
-    mock_user_store, mock_youtube_connector, mock_twitter_connector, mock_storage, mock_audio_store
+    social_handler_context_builder,
+    mock_user_store,
+    mock_youtube_connector,
+    mock_twitter_connector,
+    mock_storage,
+    mock_audio_store,
 ):
-    return {
-        "user_store": mock_user_store,
-        "youtube_connector": mock_youtube_connector,
-        "twitter_connector": mock_twitter_connector,
-        "storage": mock_storage,
-        "audio_store": mock_audio_store,
-    }
+    return social_handler_context_builder(
+        user_store=mock_user_store,
+        youtube_connector=mock_youtube_connector,
+        twitter_connector=mock_twitter_connector,
+        storage=mock_storage,
+        audio_store=mock_audio_store,
+    )
 
 
 @pytest.fixture
